Bugfix by avoiding usage of GLTFSkeleton
Fix for off-by-n error when non-bones are added to the Godot Skeleton:
Some Godot skeletons insert non-bone GLTF nodes when they are children of bones and parents of other bones.
For example, in hips -> cloth_parent -> cloth_1 ..., cloth_parent is not a GLTF bone, but is in the Godot skeleton.

Instead of getting the bone index and asking for the Godot node by name, we instead use the "resource_name" of the GLTFNode, which works because it has already been mangled.

The other implication with this change is that we no longer need to use GLTFSkeleton, which as noted by iFire, can be unreliable.
Test case: Miraikomachi.vrm